Stabilize the rng to the Sequential Insertion test
Before this patch I would see the occasional failure when running:

    $ go test -run Sequential -count 30

The failures are consistently _not_ on the extreme quantiles, as
expected given the data-structure characteristics. I guess this is
because `Add()` and `shuffle()` (which is used by `Compress()`, that
gets triggered several times for this test given the small summary
size) use the same rng source, but I'm not even remotely close to
someone who properly understands all this math.

Shouldn't be a real problem, but I'll consider adding support for
providing a random source instead of using the default one so that
at least we get no contention from multiple things asking for a random
number to the global source.