gometalinter: Error checking on tests
With this patch we now explicitly ignore most errors coming from Add()
(as there are specialized tests for it and the consistency checks
cover the error scenario) and introduces a check for potential Merge()
and Compress() errors.

Not really happy with the mass ignore, but I do like the `errcheck`
linter, so there's that.